Move private delegate extensions to runtime
